add key value store to common and aws.
