--- cloud/aws/deploy/keyvalue/dynamodb.go
@@ -1,68 +1,0 @@
-// Copyright Nitric Pty Ltd.
-//
-// SPDX-License-Identifier: Apache-2.0
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at:
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package keyvalue
-
-import (
-	"github.com/nitrictech/nitric/cloud/common/deploy/resources"
-	"github.com/nitrictech/nitric/cloud/common/deploy/tags"
-	v1 "github.com/nitrictech/nitric/core/pkg/proto/deployments/v1"
-	"github.com/pulumi/pulumi-aws/sdk/v5/go/aws/dynamodb"
-	"github.com/pulumi/pulumi/sdk/v3/go/pulumi"
-)
-
-type DynamodbKeyValueStore struct {
-	pulumi.ResourceState
-
-	Table *dynamodb.Table
-	Name  string
-}
-
-type DynamodbKeyValueStoreArgs struct {
-	StackID       string
-	KeyValueStore *v1.KeyValueStore
-}
-
-func NewDynamodbKeyValueStore(ctx *pulumi.Context, name string, args *DynamodbKeyValueStoreArgs, opts ...pulumi.ResourceOption) (*DynamodbKeyValueStore, error) {
-	res := &DynamodbKeyValueStore{Name: name}
-
-	err := ctx.RegisterComponentResource("nitric:keyvalue:Dynamodb", name, res, opts...)
-	if err != nil {
-		return nil, err
-	}
-
-	res.Table, err = dynamodb.NewTable(ctx, name, &dynamodb.TableArgs{
-		Attributes: dynamodb.TableAttributeArray{
-			&dynamodb.TableAttributeArgs{
-				Name: pulumi.String("_pk"),
-				Type: pulumi.String("S"),
-			},
-			&dynamodb.TableAttributeArgs{
-				Name: pulumi.String("_sk"),
-				Type: pulumi.String("S"),
-			},
-		},
-		HashKey:     pulumi.String("_pk"),
-		RangeKey:    pulumi.String("_sk"),
-		BillingMode: pulumi.String("PAY_PER_REQUEST"),
-		Tags:        pulumi.ToStringMap(tags.Tags(args.StackID, name, resources.Collection)),
-	})
-	if err != nil {
-		return nil, err
-	}
-
-	return res, nil
-}
